Move more functions out of misc list...
More eval-like code moved into eval.
