+++ mathics/builtin/list/rearrange.py
@@ -11,12 +11,13 @@ from itertools import chain
 from typing import Callable
 
 from mathics.builtin.base import Builtin, MessageException
-from mathics.core.atoms import Integer
+from mathics.core.atoms import Integer, Integer0
 from mathics.core.attributes import A_FLAT, A_ONE_IDENTITY, A_PROTECTED
+from mathics.core.evaluation import Evaluation
 from mathics.core.expression import Expression, structure
 from mathics.core.list import ListExpression
 from mathics.core.symbols import Atom, Symbol, SymbolTrue
-from mathics.core.systemsymbols import SymbolMap
+from mathics.core.systemsymbols import SymbolMap, SymbolSplit
 
 SymbolReverse = Symbol("Reverse")
 
@@ -39,22 +40,23 @@ def _is_sameq(same_test):
 
 
 class _FastEquivalence:
-    # models an equivalence relation through SameQ. for n distinct elements (each
-    # in its own bin), we expect to make O(n) comparisons (if the hash function
-    # does not fail us by distributing items very unevenly).
+    """
+    Models an equivalence relation using SameQ. for n distinct elements (each
+    in its own bin), we expect to make O(n) comparisons (if the hash function
+    does not fail us by distributing items very unevenly).
 
-    # IMPORTANT NOTE ON ATOM'S HASH FUNCTIONS / this code relies on this assumption:
-    #
-    # if SameQ[a, b] == true then hash(a) == hash(b)
-    #
-    # more specifically, this code bins items based on their hash code, and only if
-    # the hash code matches, is SameQ evoked.
-    #
-    # this assumption has been checked for these types: Integer, Real, Complex,
-    # String, Rational (*), Expression, Image; new atoms need proper hash functions
-    #
-    # (*) Rational values are sympy Rationals which are always held in reduced form
-    # and thus are hashed correctly (see sympy/core/number.py:Rational.__eq__()).
+    IMPORTANT NOTE ON ATOM'S HASH FUNCTIONS / this code relies on this assumption:
+       if SameQ[a, b] == true then hash(a) == hash(b)
+
+    Specifically, this code bins items based on their hash code, and only if
+    the hash code matches, is SameQ evoked.
+
+    This assumption has been checked for these types: Integer, Real, Complex,
+    String, Rational (*), Expression, Image; new atoms need proper hash functions
+
+    (*) Rational values are sympy Rationals which are always held in reduced form
+    and thus are hashed correctly (see sympy/core/number.py:Rational.__eq__()).
+    """
 
     def __init__(self):
         self._hashes = defaultdict(list)
@@ -65,6 +67,195 @@ class _FastEquivalence:
     def sameQ(self, a, b) -> bool:
         """Mathics SameQ"""
         return a.sameQ(b)
+
+
+class _IllegalPaddingDepth(Exception):
+    def __init__(self, level):
+        self.level = level
+
+
+class _Pad(Builtin):
+    messages = {
+        "normal": "Expression at position 1 in `` must not be an atom.",
+        "level": "Cannot pad list `3` which has `4` using padding `1` which specifies `2`.",
+        "ilsm": "Expected an integer or a list of integers at position `1` in `2`.",
+    }
+
+    rules = {"%(name)s[l_]": "%(name)s[l, Automatic]"}
+
+    @staticmethod
+    def _find_dims(expr):
+        def dive(expr, level):
+            if isinstance(expr, Expression):
+                if expr.elements:
+                    return max(dive(x, level + 1) for x in expr.elements)
+                else:
+                    return level + 1
+            else:
+                return level
+
+        def calc(expr, dims, level):
+            if isinstance(expr, Expression):
+                for x in expr.elements:
+                    calc(x, dims, level + 1)
+                dims[level] = max(dims[level], len(expr.elements))
+
+        dims = [0] * dive(expr, 0)
+        calc(expr, dims, 0)
+        return dims
+
+    @staticmethod
+    def _build(
+        element, n, x, m, level, mode
+    ):  # mode < 0 for left pad, > 0 for right pad
+        if not n:
+            return element
+        if not isinstance(element, Expression):
+            raise _IllegalPaddingDepth(level)
+
+        if isinstance(m, (list, tuple)):
+            current_m = m[0] if m else 0
+            next_m = m[1:]
+        else:
+            current_m = m
+            next_m = m
+
+        def clip(a, d, s):
+            assert d != 0
+            if s < 0:
+                return a[-d:]  # end with a[-1]
+            else:
+                return a[:d]  # start with a[0]
+
+        def padding(amount, sign):
+            if amount == 0:
+                return []
+            elif len(n) > 1:
+                return [
+                    _Pad._build(ListExpression(), n[1:], x, next_m, level + 1, mode)
+                ] * amount
+            else:
+                return clip(x * (1 + amount // len(x)), amount, sign)
+
+        elements = element.elements
+        d = n[0] - len(elements)
+        if d < 0:
+            new_elements = clip(elements, d, mode)
+            padding_main = []
+        elif d >= 0:
+            new_elements = elements
+            padding_main = padding(d, mode)
+
+        if current_m > 0:
+            padding_margin = padding(
+                min(current_m, len(new_elements) + len(padding_main)), -mode
+            )
+
+            if len(padding_margin) > len(padding_main):
+                padding_main = []
+                new_elements = clip(
+                    new_elements, -(len(padding_margin) - len(padding_main)), mode
+                )
+            elif len(padding_margin) > 0:
+                padding_main = clip(padding_main, -len(padding_margin), mode)
+        else:
+            padding_margin = []
+
+        if len(n) > 1:
+            new_elements = (
+                _Pad._build(e, n[1:], x, next_m, level + 1, mode) for e in new_elements
+            )
+
+        if mode < 0:
+            parts = (padding_main, new_elements, padding_margin)
+        else:
+            parts = (padding_margin, new_elements, padding_main)
+
+        return Expression(element.get_head(), *list(chain(*parts)))
+
+    def _pad(self, in_l, in_n, in_x, in_m, evaluation, expr):
+        if not isinstance(in_l, Expression):
+            evaluation.message(self.get_name(), "normal", expr())
+            return
+
+        py_n = None
+        if isinstance(in_n, Symbol) and in_n.get_name() == "System`Automatic":
+            py_n = _Pad._find_dims(in_l)
+        elif in_n.get_head_name() == "System`List":
+            if all(isinstance(element, Integer) for element in in_n.elements):
+                py_n = [element.get_int_value() for element in in_n.elements]
+        elif isinstance(in_n, Integer):
+            py_n = [in_n.get_int_value()]
+
+        if py_n is None:
+            evaluation.message(self.get_name(), "ilsm", 2, expr())
+            return
+
+        if in_x.get_head_name() == "System`List":
+            py_x = in_x.elements
+        else:
+            py_x = [in_x]
+
+        if isinstance(in_m, Integer):
+            py_m = in_m.get_int_value()
+        else:
+            if not all(isinstance(x, Integer) for x in in_m.elements):
+                evaluation.message(self.get_name(), "ilsm", 4, expr())
+                return
+            py_m = [x.get_int_value() for x in in_m.elements]
+
+        try:
+            return _Pad._build(in_l, py_n, py_x, py_m, 1, self._mode)
+        except _IllegalPaddingDepth as e:
+
+            def levels(k):
+                if k == 1:
+                    return "1 level"
+                else:
+                    return "%d levels" % k
+
+            evaluation.message(
+                self.get_name(),
+                "level",
+                in_n,
+                levels(len(py_n)),
+                in_l,
+                levels(e.level - 1),
+            )
+            return None
+
+    def eval_zero(self, element, n, evaluation):
+        "%(name)s[element_, n_]"
+        return self._pad(
+            element,
+            n,
+            Integer0,
+            Integer0,
+            evaluation,
+            lambda: Expression(self.get_name(), element, n),
+        )
+
+    def eval(self, element, n, x, evaluation):
+        "%(name)s[element_, n_, x_]"
+        return self._pad(
+            element,
+            n,
+            x,
+            Integer0,
+            evaluation,
+            lambda: Expression(self.get_name(), element, n, x),
+        )
+
+    def eval_margin(self, element, n, x, m, evaluation):
+        "%(name)s[element_, n_, x_, m_]"
+        return self._pad(
+            element,
+            n,
+            x,
+            m,
+            evaluation,
+            lambda: Expression(self.get_name(), element, n, x, m),
+        )
 
 
 class _SlowEquivalence:
@@ -117,7 +308,7 @@ class _GatherOperation(Builtin):
         ),
     }
 
-    def apply(self, values, test, evaluation):
+    def apply(self, values, test, evaluation: Evaluation):
         "%(name)s[values_, test_]"
         if not self._check_list(values, test, evaluation):
             return
@@ -129,7 +320,7 @@ class _GatherOperation(Builtin):
                 values, values, _SlowEquivalence(test, evaluation, self.get_name())
             )
 
-    def _check_list(self, values, arg2, evaluation):
+    def _check_list(self, values, arg2, evaluation: Evaluation):
         if isinstance(values, Atom):
             expr = Expression(Symbol(self.get_name()), values, arg2)
             evaluation.message(self.get_name(), "normal", 1, expr)
@@ -163,7 +354,7 @@ class _GatherOperation(Builtin):
 class _Rotate(Builtin):
     messages = {"rspec": "`` should be an integer or a list of integers."}
 
-    def _rotate(self, expr, n, evaluation):
+    def _rotate(self, expr, n, evaluation: Evaluation):
         if not isinstance(expr, Expression):
             return expr
 
@@ -181,11 +372,11 @@ class _Rotate(Builtin):
 
         return expr.restructure(expr.head, new_elements, evaluation)
 
-    def apply_one(self, expr, evaluation):
+    def apply_one(self, expr, evaluation: Evaluation):
         "%(name)s[expr_]"
         return self._rotate(expr, [1], evaluation)
 
-    def apply(self, expr, n, evaluation):
+    def apply(self, expr, n, evaluation: Evaluation):
         "%(name)s[expr_, n_]"
         if isinstance(n, Integer):
             py_cycles = [n.get_int_value()]
@@ -296,7 +487,7 @@ class Catenate(Builtin):
     summary_text = "catenate elements from a list of lists"
     messages = {"invrp": "`1` is not a list."}
 
-    def apply(self, lists, evaluation):
+    def apply(self, lists, evaluation: Evaluation):
         "Catenate[lists_List]"
 
         def parts():
@@ -424,14 +615,19 @@ class Gather(_GatherOperation):
 
 class GatherBy(_GatherOperation):
     """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/GatherBy.html</url>
+    <url>
+    :WMA link:
+    https://reference.wolfram.com/language/ref/GatherBy.html</url>
 
     <dl>
       <dt>'GatherBy[$list$, $f$]'
-      <dd>gathers elements of $list$ into sub lists of items whose image under $f$ identical.
+      <dd>gathers elements of $list$ into sub lists of items whose image \
+      under $f$ identical.
 
       <dt>'GatherBy[$list$, {$f$, $g$, ...}]'
-      <dd>gathers elements of $list$ into sub lists of items whose image under $f$ identical. Then, gathers these sub lists again into sub sub lists, that are identical under $g.
+      <dd>gathers elements of $list$ into sub lists of items whose image \
+      under $f$ identical. Then, gathers these sub lists again into sub \
+      sub lists, that are identical under $g.
     </dl>
 
     >> GatherBy[{{1, 3}, {2, 2}, {1, 1}}, Total]
@@ -455,7 +651,7 @@ class GatherBy(_GatherOperation):
     summary_text = "gather based on values of a function applied to elements"
     _bin = _GatherBin
 
-    def apply(self, values, func, evaluation):
+    def apply(self, values, func, evaluation: Evaluation):
         "%(name)s[values_, func_]"
 
         if not self._check_list(values, func, evaluation):
@@ -470,7 +666,9 @@ class GatherBy(_GatherOperation):
 
 class Join(Builtin):
     """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/Join.html</url>
+    <url>
+    :WMA link:
+    https://reference.wolfram.com/language/ref/Join.html</url>
 
     <dl>
       <dt>'Join[$l1$, $l2$]'
@@ -506,7 +704,7 @@ class Join(Builtin):
     attributes = A_FLAT | A_ONE_IDENTITY | A_PROTECTED
     summary_text = "join lists together at any level"
 
-    def apply(self, lists, evaluation):
+    def apply(self, lists, evaluation: Evaluation):
         "Join[lists___]"
 
         result = []
@@ -526,6 +724,82 @@ class Join(Builtin):
             return sequence[0].restructure(head, result, evaluation, deps=sequence)
         else:
             return ListExpression()
+
+
+class PadLeft(_Pad):
+    """
+    <url>:WMA link:https://reference.wolfram.com/language/ref/PadLeft.html</url>
+
+    <dl>
+      <dt>'PadLeft[$list$, $n$]'
+      <dd>pads $list$ to length $n$ by adding 0 on the left.
+      <dt>'PadLeft[$list$, $n$, $x$]'
+      <dd>pads $list$ to length $n$ by adding $x$ on the left.
+      <dt>'PadLeft[$list$, {$n1$, $n2, ...}, $x$]'
+      <dd>pads $list$ to lengths $n1$, $n2$ at levels 1, 2, ... respectively by adding $x$ on the left.
+      <dt>'PadLeft[$list$, $n$, $x$, $m$]'
+      <dd>pads $list$ to length $n$ by adding $x$ on the left and adding a margin of $m$ on the right.
+      <dt>'PadLeft[$list$, $n$, $x$, {$m1$, $m2$, ...}]'
+      <dd>pads $list$ to length $n$ by adding $x$ on the left and adding margins of $m1$, $m2$, ...
+         on levels 1, 2, ... on the right.
+      <dt>'PadLeft[$list$]'
+      <dd>turns the ragged list $list$ into a regular list by adding 0 on the left.
+    </dl>
+
+    >> PadLeft[{1, 2, 3}, 5]
+     = {0, 0, 1, 2, 3}
+    >> PadLeft[x[a, b, c], 5]
+     = x[0, 0, a, b, c]
+    >> PadLeft[{1, 2, 3}, 2]
+     = {2, 3}
+    >> PadLeft[{{}, {1, 2}, {1, 2, 3}}]
+     = {{0, 0, 0}, {0, 1, 2}, {1, 2, 3}}
+    >> PadLeft[{1, 2, 3}, 10, {a, b, c}, 2]
+     = {b, c, a, b, c, 1, 2, 3, a, b}
+    >> PadLeft[{{1, 2, 3}}, {5, 2}, x, 1]
+     = {{x, x}, {x, x}, {x, x}, {3, x}, {x, x}}
+    """
+
+    _mode = -1
+    summary_text = "pad out by the left a ragged array to make a matrix"
+
+
+class PadRight(_Pad):
+    """
+    <url>:WMA link:https://reference.wolfram.com/language/ref/PadRight.html</url>
+
+    <dl>
+      <dt>'PadRight[$list$, $n$]'
+      <dd>pads $list$ to length $n$ by adding 0 on the right.
+      <dt>'PadRight[$list$, $n$, $x$]'
+      <dd>pads $list$ to length $n$ by adding $x$ on the right.
+      <dt>'PadRight[$list$, {$n1$, $n2, ...}, $x$]'
+      <dd>pads $list$ to lengths $n1$, $n2$ at levels 1, 2, ... respectively by adding $x$ on the right.
+      <dt>'PadRight[$list$, $n$, $x$, $m$]'
+      <dd>pads $list$ to length $n$ by adding $x$ on the left and adding a margin of $m$ on the left.
+      <dt>'PadRight[$list$, $n$, $x$, {$m1$, $m2$, ...}]'
+      <dd>pads $list$ to length $n$ by adding $x$ on the right and adding margins of $m1$, $m2$, ...
+         on levels 1, 2, ... on the left.
+      <dt>'PadRight[$list$]'
+      <dd>turns the ragged list $list$ into a regular list by adding 0 on the right.
+    </dl>
+
+    >> PadRight[{1, 2, 3}, 5]
+     = {1, 2, 3, 0, 0}
+    >> PadRight[x[a, b, c], 5]
+     = x[a, b, c, 0, 0]
+    >> PadRight[{1, 2, 3}, 2]
+     = {1, 2}
+    >> PadRight[{{}, {1, 2}, {1, 2, 3}}]
+     = {{0, 0, 0}, {1, 2, 0}, {1, 2, 3}}
+    >> PadRight[{1, 2, 3}, 10, {a, b, c}, 2]
+     = {b, c, 1, 2, 3, a, b, c, a, b}
+    >> PadRight[{{1, 2, 3}}, {5, 2}, x, 1]
+     = {{x, x}, {x, 1}, {x, x}, {x, x}, {x, x}}
+    """
+
+    _mode = 1
+    summary_text = "pad out by the right a ragged array to make a matrix"
 
 
 class Partition(Builtin):
@@ -560,7 +834,7 @@ class Partition(Builtin):
         "Parition[list_, n_, d_, k]": "Partition[list, n, d, {k, k}]",
     }
 
-    def _partition(self, expr, n, d, evaluation):
+    def _partition(self, expr, n, d, evaluation: Evaluation):
         assert n > 0 and d > 0
 
         inner = structure("List", expr, evaluation)
@@ -581,12 +855,12 @@ class Partition(Builtin):
 
         return outer(slices())
 
-    def apply_no_overlap(self, li, n, evaluation):
+    def apply_no_overlap(self, li, n, evaluation: Evaluation):
         "Partition[li_List, n_Integer]"
         # TODO: Error checking
         return self._partition(li, n.get_int_value(), n.get_int_value(), evaluation)
 
-    def apply(self, li, n, d, evaluation):
+    def apply(self, li, n, d, evaluation: Evaluation):
         "Partition[li_List, n_Integer, d_Integer]"
         # TODO: Error checking
         return self._partition(li, n.get_int_value(), d.get_int_value(), evaluation)
@@ -655,11 +929,11 @@ class Reverse(Builtin):
 
         return expr
 
-    def apply_top_level(self, expr, evaluation):
+    def apply_top_level(self, expr, evaluation: Evaluation):
         "Reverse[expr_]"
         return Reverse._reverse(expr, 1, (1,), evaluation)
 
-    def apply(self, expr, levels, evaluation):
+    def apply(self, expr, levels, evaluation: Evaluation):
         "Reverse[expr_, levels_]"
         if isinstance(levels, Integer):
             py_levels = [levels.get_int_value()]
@@ -733,7 +1007,7 @@ class Riffle(Builtin):
 
     summary_text = "intersperse additional elements"
 
-    def apply(self, list, sep, evaluation):
+    def apply(self, list, sep, evaluation: Evaluation):
         "Riffle[list_List, sep_]"
 
         if sep.has_form("List", None):
@@ -802,6 +1076,150 @@ class RotateRight(_Rotate):
     _sign = -1
 
 
+class Split(Builtin):
+    """
+    <url>:WMA link:https://reference.wolfram.com/language/ref/Split.html</url>
+
+    <dl>
+      <dt>'Split[$list$]'
+      <dd>splits $list$ into collections of consecutive identical elements.
+      <dt>'Split[$list$, $test$]'
+      <dd>splits $list$ based on whether the function $test$ yields
+        'True' on consecutive elements.
+    </dl>
+
+    >> Split[{x, x, x, y, x, y, y, z}]
+     = {{x, x, x}, {y}, {x}, {y, y}, {z}}
+
+    #> Split[{x, x, x, y, x, y, y, z}, x]
+     = {{x}, {x}, {x}, {y}, {x}, {y}, {y}, {z}}
+
+    Split into increasing or decreasing runs of elements
+    >> Split[{1, 5, 6, 3, 6, 1, 6, 3, 4, 5, 4}, Less]
+     = {{1, 5, 6}, {3, 6}, {1, 6}, {3, 4, 5}, {4}}
+
+    >> Split[{1, 5, 6, 3, 6, 1, 6, 3, 4, 5, 4}, Greater]
+     = {{1}, {5}, {6, 3}, {6, 1}, {6, 3}, {4}, {5, 4}}
+
+    Split based on first element
+    >> Split[{x -> a, x -> y, 2 -> a, z -> c, z -> a}, First[#1] === First[#2] &]
+     = {{x -> a, x -> y}, {2 -> a}, {z -> c, z -> a}}
+
+    #> Split[{}]
+     = {}
+
+    #> A[x__] := 321 /; Length[{x}] == 5;
+    #> Split[A[x, x, x, y, x, y, y, z]]
+     = 321
+    #> ClearAll[A];
+    """
+
+    rules = {
+        "Split[list_]": "Split[list, SameQ]",
+    }
+
+    messages = {
+        "normal": "Nonatomic expression expected at position `1` in `2`.",
+    }
+    summary_text = "split into runs of identical elements"
+
+    def eval(self, mlist, test, evaluation: Evaluation):
+        "Split[mlist_, test_]"
+
+        expr = Expression(SymbolSplit, mlist, test)
+
+        if isinstance(mlist, Atom):
+            evaluation.message("Select", "normal", 1, expr)
+            return
+
+        if not mlist.elements:
+            return Expression(mlist.head)
+
+        result = [[mlist.elements[0]]]
+        for element in mlist.elements[1:]:
+            applytest = Expression(test, result[-1][-1], element)
+            if applytest.evaluate(evaluation) is SymbolTrue:
+                result[-1].append(element)
+            else:
+                result.append([element])
+
+        inner = structure("List", mlist, evaluation)
+        outer = structure(mlist.head, inner, evaluation)
+        return outer([inner(t) for t in result])
+
+
+class SplitBy(Builtin):
+    """
+    <url>
+    :WMA link:
+    https://reference.wolfram.com/language/ref/SplitBy.html</url>
+
+    <dl>
+      <dt>'SplitBy[$list$, $f$]'
+      <dd>splits $list$ into collections of consecutive elements
+        that give the same result when $f$ is applied.
+    </dl>
+
+    >> SplitBy[Range[1, 3, 1/3], Round]
+     = {{1, 4 / 3}, {5 / 3, 2, 7 / 3}, {8 / 3, 3}}
+
+    >> SplitBy[{1, 2, 1, 1.2}, {Round, Identity}]
+     = {{{1}}, {{2}}, {{1}, {1.2}}}
+
+    #> SplitBy[Tuples[{1, 2}, 3], First]
+     = {{{1, 1, 1}, {1, 1, 2}, {1, 2, 1}, {1, 2, 2}}, {{2, 1, 1}, {2, 1, 2}, {2, 2, 1}, {2, 2, 2}}}
+    """
+
+    messages = {
+        "normal": "Nonatomic expression expected at position `1` in `2`.",
+    }
+
+    rules = {
+        "SplitBy[list_]": "SplitBy[list, Identity]",
+    }
+
+    summary_text = "split based on values of a function applied to elements"
+
+    def eval(self, mlist, func, evaluation: Evaluation):
+        "SplitBy[mlist_, func_?NotListQ]"
+
+        expr = Expression(SymbolSplit, mlist, func)
+
+        if isinstance(mlist, Atom):
+            evaluation.message("Select", "normal", 1, expr)
+            return
+
+        plist = [t for t in mlist.elements]
+
+        result = [[plist[0]]]
+        prev = Expression(func, plist[0]).evaluate(evaluation)
+        for element in plist[1:]:
+            curr = Expression(func, element).evaluate(evaluation)
+            if curr == prev:
+                result[-1].append(element)
+            else:
+                result.append([element])
+            prev = curr
+
+        inner = structure("List", mlist, evaluation)
+        outer = structure(mlist.head, inner, evaluation)
+        return outer([inner(t) for t in result])
+
+    def eval_multiple(self, mlist, funcs, evaluation: Evaluation):
+        "SplitBy[mlist_, funcs_List]"
+        expr = Expression(SymbolSplit, mlist, funcs)
+
+        if isinstance(mlist, Atom):
+            evaluation.message("Select", "normal", 1, expr)
+            return
+
+        result = mlist
+        for f in funcs.elements[::-1]:
+            result = self.eval(result, f, evaluation)
+
+        return result
+
+
 class Tally(_GatherOperation):
     """
     <url>:WMA link:https://reference.wolfram.com/language/ref/Tally.html</url>
@@ -867,11 +1285,14 @@ class Union(_SetOperation):
 
 class Intersection(_SetOperation):
     """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/Intersection.html</url>
+    <url>
+    :WMA link:
+    https://reference.wolfram.com/language/ref/Intersection.html</url>
 
     <dl>
       <dt>'Intersection[$a$, $b$, ...]'
-      <dd>gives the intersection of the sets. The resulting list will be sorted and each element will only occur once.
+      <dd>gives the intersection of the sets. The resulting list \
+      will be sorted and each element will only occur once.
     </dl>
 
     >> Intersection[{1000, 100, 10, 1}, {1, 5, 10, 15}]
