+++ mathics/core/systemsymbols.py
@@ -44,6 +44,7 @@ SymbolC = Symbol("System`C")
 SymbolCatalan = Symbol("System`Cases")
 SymbolCatalan = Symbol("System`Catalan")
 SymbolCeiling = Symbol("System`Ceiling")
+SymbolClusteringComponents = Symbol("System`ClusteringComponents")
 SymbolColorConvert = Symbol("System`ColorConvert")
 SymbolColorData = Symbol("System`ColorData")
 SymbolCompile = Symbol("System`Compile")
@@ -53,6 +54,7 @@ SymbolComplexInfinity = Symbol("System`ComplexInfinity")
 SymbolCondition = Symbol("System`Condition")
 SymbolConditionalExpression = Symbol("System`ConditionalExpression")
 SymbolConjugate = Symbol("System`Conjugate")
+SymbolContainsOnly = Symbol("System`ContainsOnly")
 SymbolContext = Symbol("System`$Context")
 SymbolContextPath = Symbol("System`$ContextPath")
 SymbolContinue = Symbol("System`Continue")
@@ -78,6 +80,7 @@ SymbolExportString = Symbol("System`ExportString")
 SymbolFaceForm = Symbol("System`FaceForm")
 SymbolFactorial = Symbol("System`Factorial")
 SymbolFailed = Symbol("System`$Failed")
+SymbolFindClusters = Symbol("System`FindClusters")
 SymbolFloor = Symbol("System`Floor")
 SymbolFormat = Symbol("System`Format")
 SymbolFractionBox = Symbol("System`FractionBox")
@@ -188,6 +191,7 @@ SymbolSimplify = Symbol("System`Simplify")
 SymbolSin = Symbol("System`Sin")
 SymbolSinh = Symbol("System`Sinh")
 SymbolSlot = Symbol("System`Slot")
+SymbolSplit = Symbol("System`Split")
 SymbolSqrt = Symbol("System'Sqrt")
 SymbolSqrtBox = Symbol("System`SqrtBox")
 SymbolStandardForm = Symbol("System`StandardForm")
